requirements.txt
efficientnet==1.1.0
ipython==7.11.1
Keras==2.3.1
Keras_Applications==1.0.8
numpy==1.16.6
opencv_contrib_python==4.2.0.32
opencv_python==4.1.2.30
scikit_learn==0.22.1
six==1.13.0
tensorflow_gpu==1.15.0
